refactor: page API 분리
